+++ src/server/js/index.js
@@ -5,3 +5,5 @@ import '../../common/js/index.js';
 
 import './check_npms.js';
 import './methods.js';
+import './publish.js';
+import './validate.js';