+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - Upgrade pwix:layout version requirement to get ui_layout.less constants
     - Define mandatoryFieldsBorder both as a configuration option and a acUserLogin option (bumping candidate version number)
     - Add a dependency on pwix:core-ui
+    - Validate the user account structure before creation on server side
 
 ### 1.2.5